requirements.txt
gunicorn
flask
keras
numpy
Pillow
tensorflow